Edited original code to work on mac.

+++ camera.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
+#include "glm.hpp"
+#include "glm/gtc/matrix_transform.hpp"
 #include <iostream>
 
 enum Direction {
@@ -20,7 +20,7 @@ public:
 
 	double pitch;
 	double yaw;
-	double fov;
+	float fov;
 	double sensitivity;
 	double speed;
